+++ model_pb_announce_item.go
@@ -8,17 +8,18 @@
  */
 
 package rum_sdk
+
 // PbAnnounceItem struct for PbAnnounceItem
 type PbAnnounceItem struct {
-	Action int32 `json:"Action,omitempty"`
+	Action             int64  `json:"Action,omitempty"`
 	AnnouncerSignature string `json:"AnnouncerSignature,omitempty"`
-	EncryptPubkey string `json:"EncryptPubkey,omitempty"`
-	GroupId string `json:"GroupId,omitempty"`
-	Memo string `json:"Memo,omitempty"`
-	OwnerPubkey string `json:"OwnerPubkey,omitempty"`
-	OwnerSignature string `json:"OwnerSignature,omitempty"`
-	Result int32 `json:"Result,omitempty"`
-	SignPubkey string `json:"SignPubkey,omitempty"`
-	TimeStamp string `json:"TimeStamp,omitempty"`
-	Type int32 `json:"Type,omitempty"`
+	EncryptPubkey      string `json:"EncryptPubkey,omitempty"`
+	GroupId            string `json:"GroupId,omitempty"`
+	Memo               string `json:"Memo,omitempty"`
+	OwnerPubkey        string `json:"OwnerPubkey,omitempty"`
+	OwnerSignature     string `json:"OwnerSignature,omitempty"`
+	Result             int64  `json:"Result,omitempty"`
+	SignPubkey         string `json:"SignPubkey,omitempty"`
+	TimeStamp          string `json:"TimeStamp,omitempty"`
+	Type               int64  `json:"Type,omitempty"`
 }
